Move filter-related code from azul.service to azul.filters
