+++ .mypy.ini
@@ -60,6 +60,7 @@ modules =
     azul.plugins.metadata.hca.service.filter,
     azul.service.storage_service,
     azul.service.source_service,
+    azul.filters,
     azul.service,
     azul.service.source_controller,
     azul.service.query_controller,
